+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 v0.6
 ---------
 - replaced unicode icons with SVG icons, because unicode got colored in Chrome
-- added highlighting for duplicated meta and title tags
+- added highlighting for duplicate meta, base and title tags
 - changed styles to force break values if too long and additionally truncate them (e.g. Base64 encoded values)
 
 v0.5